pyproject.toml
[tool.poetry]
name = "app"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.8"
fastapi = {extras = ["full"], version = "^0.63.0"}
SQLAlchemy = "1.3"
python-jose = "^3.2.0"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
uvicorn = "^0.13.4"
alembic = "^1.5.7"
psycopg2-binary = "^2.8.6"
pydantic = {extras = ["email"], version = "^1.8.1"}
python-multipart = "^0.0.5"
fastapi-mail = "^0.3.5"
pytest-asyncio = "^0.15.1"
aioredis = "2.0.0a1"
Pillow = "^8.2.0"
aiofiles = "^0.6.0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
pylint = "^2.7.2"
pytest-cov = "^2.11.1"
pytest-html = "^3.1.1"
autopep8 = "^1.5.6"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
